MEXC-Inverse-btc-trader/requirements.txt
requests>=2.25.1
asyncpg==0.29.0
python-dotenv==1.0.1
structlog==23.3.0
websockets==12.0
aiohttp==3.9.3
pandas==2.2.3

# For LSTM & ML
numpy==1.24.2
tensorflow==2.15.0
scikit-learn==1.2.2
redis==4.5.1

# Technical indicators
ta==0.10.1

# For hyperparameter tuning and feature importance analysis
keras-tuner>=1.0.3
shap>=0.41.0
